pyproject.toml
[tool.poetry]
name = "crypto"
version = "0.1.0"
description = ""
authors = ["Matheus Cardoso <matheus.cardoso@enline-transmission.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.2"
prophet = "^1.1.2"
requests = "^2.28.2"
pandas = "^1.5.3"

[[tool.poetry.source]]
name = "stanfordnlp"
url = "https://stanfordnlp.github.io/stanfordnlp_pytorch/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
